Implemented RemoveTell! Closes #11
Improved the help text of MM and Welch, and added list pastebin-ing to Welch.
Possibly fixed a small bug with SentTells, if a user changed the case of letters in their name.
Maybe got rid of the dependency on the mono forms library. Will have to test when next on Ubuntu.