Changed Event to use DateTimeOffset objects for parsing instead of plain old DateTime objects, per dave_random's request.
Took the time to invert a bunch of the logic in the various functions under countdown, so they are indented less, and exit conditions are more obvious.
Also fixed TimeTill and TimeSince to use whatever Desert Bus events are closest, rather than hard-coded DB5 and DB4. All Desert Bus events should be changed to the format 'DB# Started' and 'DB# Ended', however.